feat(address-service): DOMA-4064 Plugins refactoring
- Extract function for saving address
- Remove many-to-many relation
- abstract classes moved to folders
- stop returning address sources from service
- remove field `addressSources` from plugin
