+++ apps/address-service/domains/address/schema/Address.js
@@ -2,7 +2,7 @@
  * Generated by `createschema address.Address 'source:Text; address:Text; key:Text; meta:Json'`
  */
 
-const { Text, Relationship } = require('@keystonejs/fields')
+const { Text } = require('@keystonejs/fields')
 const { Json } = require('@open-condo/keystone/fields')
 const { GQLListSchema } = require('@open-condo/keystone/schema')
 const { historical, versioned, uuided, tracked, softDeleted } = require('@open-condo/keystone/plugins')
@@ -13,14 +13,6 @@ const Address = new GQLListSchema('Address', {
     schemaDoc: 'A model containing data on the particular building\'s address',
     labelResolver: ({ address }) => address,
     fields: {
-
-        sources: {
-            schemaDoc: 'The origin which used to detect the address',
-            type: Relationship,
-            ref: 'AddressSource.address',
-            many: true,
-            isRequired: true,
-        },
 
         address: {
             schemaDoc: 'The normalized address itself in one string',
