+++ apps/address-service/domains/common/utils/services/search/plugins/AbstractSearchPlugin.js
