+++ apps/address-service/domains/common/utils/services/search/providers/AbstractSearchProvider.js
