+++ apps/condo/schema.ts
@@ -33852,8 +33852,6 @@ export type Property = {
   id: Scalars['ID'];
   /**  Normalized address  */
   address?: Maybe<Scalars['String']>;
-  /**  The origins of the address (some strings which may looks like real address or some id)  */
-  addressSources?: Maybe<Scalars['JSON']>;
   /**  The unique key of the address  */
   addressKey?: Maybe<Scalars['String']>;
   /**  Property address components  */
@@ -33884,7 +33882,6 @@ export type PropertyCreateInput = {
   yearOfConstruction?: Maybe<Scalars['String']>;
   area?: Maybe<Scalars['String']>;
   address?: Maybe<Scalars['String']>;
-  addressSources?: Maybe<Scalars['JSON']>;
   addressKey?: Maybe<Scalars['String']>;
   addressMeta?: Maybe<Scalars['JSON']>;
   v?: Maybe<Scalars['Int']>;
@@ -33922,7 +33919,6 @@ export type PropertyHistoryRecord = {
   area?: Maybe<Scalars['String']>;
   id: Scalars['ID'];
   address?: Maybe<Scalars['String']>;
-  addressSources?: Maybe<Scalars['JSON']>;
   addressKey?: Maybe<Scalars['String']>;
   addressMeta?: Maybe<Scalars['JSON']>;
   v?: Maybe<Scalars['Int']>;
@@ -33952,7 +33948,6 @@ export type PropertyHistoryRecordCreateInput = {
   yearOfConstruction?: Maybe<Scalars['String']>;
   area?: Maybe<Scalars['String']>;
   address?: Maybe<Scalars['String']>;
-  addressSources?: Maybe<Scalars['JSON']>;
   addressKey?: Maybe<Scalars['String']>;
   addressMeta?: Maybe<Scalars['JSON']>;
   v?: Maybe<Scalars['Int']>;
@@ -33988,7 +33983,6 @@ export type PropertyHistoryRecordUpdateInput = {
   yearOfConstruction?: Maybe<Scalars['String']>;
   area?: Maybe<Scalars['String']>;
   address?: Maybe<Scalars['String']>;
-  addressSources?: Maybe<Scalars['JSON']>;
   addressKey?: Maybe<Scalars['String']>;
   addressMeta?: Maybe<Scalars['JSON']>;
   v?: Maybe<Scalars['Int']>;
@@ -34116,10 +34110,6 @@ export type PropertyHistoryRecordWhereInput = {
   address_not_ends_with_i?: Maybe<Scalars['String']>;
   address_in?: Maybe<Array<Maybe<Scalars['String']>>>;
   address_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  addressSources?: Maybe<Scalars['JSON']>;
-  addressSources_not?: Maybe<Scalars['JSON']>;
-  addressSources_in?: Maybe<Array<Maybe<Scalars['JSON']>>>;
-  addressSources_not_in?: Maybe<Array<Maybe<Scalars['JSON']>>>;
   addressKey?: Maybe<Scalars['String']>;
   addressKey_not?: Maybe<Scalars['String']>;
   addressKey_contains?: Maybe<Scalars['String']>;
@@ -34261,7 +34251,6 @@ export type PropertyUpdateInput = {
   yearOfConstruction?: Maybe<Scalars['String']>;
   area?: Maybe<Scalars['String']>;
   address?: Maybe<Scalars['String']>;
-  addressSources?: Maybe<Scalars['JSON']>;
   addressKey?: Maybe<Scalars['String']>;
   addressMeta?: Maybe<Scalars['JSON']>;
   v?: Maybe<Scalars['Int']>;
@@ -34362,10 +34351,6 @@ export type PropertyWhereInput = {
   address_not_ends_with_i?: Maybe<Scalars['String']>;
   address_in?: Maybe<Array<Maybe<Scalars['String']>>>;
   address_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  addressSources?: Maybe<Scalars['JSON']>;
-  addressSources_not?: Maybe<Scalars['JSON']>;
-  addressSources_in?: Maybe<Array<Maybe<Scalars['JSON']>>>;
-  addressSources_not_in?: Maybe<Array<Maybe<Scalars['JSON']>>>;
   addressKey?: Maybe<Scalars['String']>;
   addressKey_not?: Maybe<Scalars['String']>;
   addressKey_contains?: Maybe<Scalars['String']>;
@@ -40542,8 +40527,6 @@ export type Resident = {
   id: Scalars['ID'];
   /**  Normalized address  */
   address?: Maybe<Scalars['String']>;
-  /**  The origins of the address (some strings which may looks like real address or some id)  */
-  addressSources?: Maybe<Scalars['JSON']>;
   /**  The unique key of the address  */
   addressKey?: Maybe<Scalars['String']>;
   /**  Property address components  */
@@ -40658,7 +40641,6 @@ export type ResidentCreateInput = {
   unitType?: Maybe<Scalars['String']>;
   organization?: Maybe<OrganizationRelateToOneInput>;
   address?: Maybe<Scalars['String']>;
-  addressSources?: Maybe<Scalars['JSON']>;
   addressKey?: Maybe<Scalars['String']>;
   addressMeta?: Maybe<Scalars['JSON']>;
   v?: Maybe<Scalars['Int']>;
@@ -40694,7 +40676,6 @@ export type ResidentHistoryRecord = {
   organization?: Maybe<Scalars['String']>;
   id: Scalars['ID'];
   address?: Maybe<Scalars['String']>;
-  addressSources?: Maybe<Scalars['JSON']>;
   addressKey?: Maybe<Scalars['String']>;
   addressMeta?: Maybe<Scalars['JSON']>;
   v?: Maybe<Scalars['Int']>;
@@ -40722,7 +40703,6 @@ export type ResidentHistoryRecordCreateInput = {
   unitType?: Maybe<Scalars['String']>;
   organization?: Maybe<Scalars['String']>;
   address?: Maybe<Scalars['String']>;
-  addressSources?: Maybe<Scalars['JSON']>;
   addressKey?: Maybe<Scalars['String']>;
   addressMeta?: Maybe<Scalars['JSON']>;
   v?: Maybe<Scalars['Int']>;
@@ -40756,7 +40736,6 @@ export type ResidentHistoryRecordUpdateInput = {
   unitType?: Maybe<Scalars['String']>;
   organization?: Maybe<Scalars['String']>;
   address?: Maybe<Scalars['String']>;
-  addressSources?: Maybe<Scalars['JSON']>;
   addressKey?: Maybe<Scalars['String']>;
   addressMeta?: Maybe<Scalars['JSON']>;
   v?: Maybe<Scalars['Int']>;
@@ -40862,10 +40841,6 @@ export type ResidentHistoryRecordWhereInput = {
   address_not_ends_with_i?: Maybe<Scalars['String']>;
   address_in?: Maybe<Array<Maybe<Scalars['String']>>>;
   address_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  addressSources?: Maybe<Scalars['JSON']>;
-  addressSources_not?: Maybe<Scalars['JSON']>;
-  addressSources_in?: Maybe<Array<Maybe<Scalars['JSON']>>>;
-  addressSources_not_in?: Maybe<Array<Maybe<Scalars['JSON']>>>;
   addressKey?: Maybe<Scalars['String']>;
   addressKey_not?: Maybe<Scalars['String']>;
   addressKey_contains?: Maybe<Scalars['String']>;
@@ -41143,7 +41118,6 @@ export type ResidentUpdateInput = {
   unitType?: Maybe<Scalars['String']>;
   organization?: Maybe<OrganizationRelateToOneInput>;
   address?: Maybe<Scalars['String']>;
-  addressSources?: Maybe<Scalars['JSON']>;
   addressKey?: Maybe<Scalars['String']>;
   addressMeta?: Maybe<Scalars['JSON']>;
   v?: Maybe<Scalars['Int']>;
@@ -41210,10 +41184,6 @@ export type ResidentWhereInput = {
   address_not_ends_with_i?: Maybe<Scalars['String']>;
   address_in?: Maybe<Array<Maybe<Scalars['String']>>>;
   address_not_in?: Maybe<Array<Maybe<Scalars['String']>>>;
-  addressSources?: Maybe<Scalars['JSON']>;
-  addressSources_not?: Maybe<Scalars['JSON']>;
-  addressSources_in?: Maybe<Array<Maybe<Scalars['JSON']>>>;
-  addressSources_not_in?: Maybe<Array<Maybe<Scalars['JSON']>>>;
   addressKey?: Maybe<Scalars['String']>;
   addressKey_not?: Maybe<Scalars['String']>;
   addressKey_contains?: Maybe<Scalars['String']>;
